auth_ms/requirements.txt
cffi==1.14.6
cryptography==35.0.0
Django==3.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
pycparser==2.20
PyJWT==2.1.0
pytz==2021.3
sqlparse==0.4.2
whitenoise==5.3.0
django-heroku==0.3.1
gunicorn==19.9.0
psycopg2-binary==2.9.1
urllib3==1.23